requirements.txt
asgiref==3.2.3
attrs==19.3.0
awsebcli==3.17.0
blessed==1.17.0
boto3==1.11.9
botocore==1.13.50
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.5.0
Django==2.1.1
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-storages==1.8
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
gunicorn==20.0.4
idna==2.7
importlib-metadata==1.5.0
jmespath==0.9.4
jsonschema==2.6.0
more-itertools==8.2.0
numpy==1.18.1
packaging==20.1
pathspec==0.5.9
Pillow==7.0.0
pluggy==0.13.1
psycopg2==2.7.5
psycopg2-binary==2.8.4
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.0
pytz==2019.3
PyYAML==3.13
requests==2.20.1
s3transfer==0.3.2
scipy==1.4.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.3
wcwidth==0.1.8
websocket-client==0.57.0
whitenoise==5.0.1
zipp==2.1.0
